+++ kurtosis-module/impl/participant_network/mev_boost/mev_boost_context.go
@@ -1,0 +1,14 @@
+package mev_boost
+
+import (
+	"fmt"
+)
+
+type MEVBoostContext struct {
+	privateIPAddress string
+	port             uint16
+}
+
+func (ctx *MEVBoostContext) Endpoint() string {
+	return fmt.Sprintf("http://%s:%d", ctx.privateIPAddress, ctx.port)
+}